Added forecast for 5 days Removed a lot of unnecessary code, default pages etc Refactored some models

+++ Weather/Weather/Models/WeatherConditionsBase.cs
@@ -7,9 +7,9 @@ namespace Weather.Models
 {
     public abstract class WeatherConditionsBase
     {
-        public double Temperature { get; set; } // degrees
-        public double MinTemperature { get; set; }
-        public double MaxTemperature { get; set; }
+        public int Temperature { get; set; } // degrees
+        public int MinTemperature { get; set; }
+        public int MaxTemperature { get; set; }
 
         public string Conditions { get; set; } // description
         public string WeatherCode { get; set; } // the source for the icon representing the weather
